Clean up pylint and utilities
